Refactor BSM implementations s.t bsm is a package.
