+++ sequence/components/bsm/time_bin_bsm.py
@@ -1,0 +1,122 @@
+from typing import Any
+
+from .base import BSM
+from ..detector.base import Detector
+from ..photon import Photon
+from ...kernel.event import Event
+from ...kernel.process import Process
+from ...utils.encoding import time_bin
+
+
+class TimeBinBSM(BSM):
+    """Class modeling a time bin BSM device.
+
+    Measures incoming photons according to time bins and manages entanglement.
+
+    Attributes:
+        name (str): label for BSM instance
+        timeline (Timeline): timeline for simulation
+        detectors (list[Detector]): list of attached photon detection devices
+    """
+
+    def __init__(self, name, timeline, phase_error=0, detectors=None):
+        """Constructor for the time bin BSM class.
+
+        Args:
+            name (str): name of the beamsplitter instance.
+            timeline (Timeline): simulation timeline.
+            phase_error (float): phase error applied to polarization qubits (unused) (default 0).
+            detectors (list[dict]): list of parameters for attached detectors,
+                in dictionary format (must be of length 2) (default None).
+        """
+
+        super().__init__(name, timeline, phase_error, detectors)
+        self.encoding = "time_bin"
+        self.encoding_type = time_bin
+        self.last_res = [-1, -1]
+        assert len(self.detectors) == 2
+
+    def get(self, photon, **kwargs):
+        """See base class.
+
+        This method adds additional side effects not present in the base class.
+
+        Side Effects:
+            May call get method of one or more attached detector(s).
+            May alter the quantum state of photon and any stored photons.
+        """
+
+        super().get(photon)
+
+        if len(self.photons) != 2:
+            return
+
+        if self.get_generator().random() < self.phase_error:
+            self.photons[1].apply_phase_error()
+        # entangle photons to measure
+        self.photons[0].combine_state(self.photons[1])
+
+        # measure in bell basis
+        res = Photon.measure_multiple(self.bell_basis, self.photons, self.get_generator())
+
+        # check if we've measured as Phi+ or Phi-; these cannot be measured by the BSM
+        if res == 0 or res == 1:
+            return
+
+        early_time = self.timeline.now()
+        late_time = early_time + self.encoding_type["bin_separation"]
+
+        # measured as Psi+
+        # send both photons to the same detector at the early and late time
+        if res == 2:
+            detector_num = self.get_generator().choice([0, 1])
+
+            process = Process(self.detectors[detector_num], "get", [])
+            event = Event(int(round(early_time)), process)
+            self.timeline.schedule(event)
+            process = Process(self.detectors[detector_num], "get", [])
+            event = Event(int(round(late_time)), process)
+            self.timeline.schedule(event)
+
+        # measured as Psi-
+        # send photons to different detectors at the early and late time
+        elif res == 3:
+            detector_num = self.get_generator().choice([0, 1])
+
+            process = Process(self.detectors[detector_num], "get", [])
+            event = Event(int(round(early_time)), process)
+            self.timeline.schedule(event)
+            process = Process(self.detectors[1 - detector_num], "get", [])
+            event = Event(int(round(late_time)), process)
+            self.timeline.schedule(event)
+
+        # invalid result from measurement
+        else:
+            raise Exception("Invalid result from photon.measure_multiple")
+
+    def trigger(self, detector: Detector, info: dict[str, Any]):
+        """See base class.
+
+        This method adds additional side effects not present in the base class.
+
+        Side Effects:
+            May send a further message to any attached entities.
+        """
+
+        detector_num = self.detectors.index(detector)
+        time = info["time"]
+
+        # check if valid time
+        if round((time - self.last_res[0]) / self.encoding_type["bin_separation"]) == 1:
+            # if time - self.last_res[0] < self.resolution + self.encoding_type["bin_separation"]:
+            # pop result message
+            # Psi+
+            if detector_num == self.last_res[1]:
+                info = {'entity': 'BSM', 'info_type': 'BSM_res', 'res': 0, 'time': time}
+                self.notify(info)
+            # Psi-
+            else:
+                info = {'entity': 'BSM', 'info_type': 'BSM_res', 'res': 1, 'time': time}
+                self.notify(info)
+
+        self.last_res = [time, detector_num]
